Start to extract common Rest provider code
There is still duplication left, due to S3's REST API's behavior:

* No Last-Modified header on the response for a PUT request
* 200 on a DELETE request to an object that does not exist
